Refactor to use custom canvas tile base
